修改csv檔案上傳
SQL檔案 預存程式修改
資料表增加序號
如果已經建立資料表，則先刪除原本的資料表。
==============
上傳檔案後端部分
將別資料夾上傳的路徑複製一份到File資料夾內，並且把原本檔名改為FA61，如果FA61.csv已存在，則先刪除檔案在建立。
尚未判斷檔名有空白的問題，以及如果是空路徑上傳的判斷。

+++ csv_upload_01.aspx.cs
@@ -10,11 +10,12 @@ using System.Web.UI.WebControls;
 using System.Web.Configuration;
 using System.Data.OleDb;
 using System.IO;
+using System.Text.RegularExpressions;
 
 public partial class csv_upload_01 : System.Web.UI.Page
 {
     DataTable sdt = new DataTable();
-    string csvPath, csvFolder, csvFile, csvFilename;
+    string csvPath, csvPathT, csvFileT, csvFolder, csvFile, csvFilename;
     DataTable dt = new DataTable();
     string connString, connString1;
     int key = 0;
@@ -30,7 +31,11 @@ public partial class csv_upload_01 : System.Web.UI.Page
         csvPath = csvFolder + csvFile;
         csvFilename = csvFolder + "FA61.csv";
         //csvPath = Server.MapPath("~/Files/") + Path.GetFileName(FileUpload1.PostedFile.FileName);
-        File.Delete(csvFilename);
+        if (File.Exists(csvFilename))
+        {
+            File.Delete(csvFilename);
+        }
+
         FileUpload1.SaveAs(csvPath);
         File.Move(csvPath, csvFilename);
 
@@ -45,9 +50,9 @@ public partial class csv_upload_01 : System.Web.UI.Page
                 da.Fill(dt);
                 GridView1.DataSource = dt;
                 GridView1.DataBind();
-                
+
                 sdt = dt;
-                
+
 
                 dt.Dispose();
                 da.Dispose();
